LU-13009 wbc: add sync(2) and fsync(2) support for WBC
This patch adds sync(2) support for WBC with all three flush
modes.
For both WBC_FLUSH_LAZY and WBC_FLUSH_AGING_DROP flush modes,
sync(2) causes all pedning modifications to filesystem metadata
and cached file data to be written. All cached files under the
protection of the root WBC EX locks are flushed, and at the
meanwhile the corresponding root WBC EX locks are dropped.
For the WBC_FLUSH_AGING_KEEP flush mode, it is similar to other
two flush modes, but the root WBC EX locks are usually kept.

This patch also adds fsync(2) support for WBC_FLUSH_AGING_KEEP
flush mode.
A fsync(2) call on a file or directory must ensure tha all changes
to the file or directory, at the time of fsync(2) was called, are
flushed to the persistent storage. As the WBC subtree is cached on
the client-side MemFS, may not be flushed to the server. To ensure
the file and directory is durably stored in the persistent storage
on the server, it must ensure that all its ancestral directory
must be flushed as well when fsync(2) was called.
To achieve this goal, it needs to reverse backtracking to its
first ancestral directory that is flushed to the server (in the
state Sync(S)). And then flush and sync from the topmost ancestral
directory in Sync(S) state to the file or directory initiated the
fsync(2) call in the top-down order.
In the WBC_FLUSH_AGING_KEEP flush mode, as the root WBC EX lock is
kept during flushing, thus the open()/close() calls can be executed
locally. But in the WBC_FLUSH_LAZY and WBC_FLUSH_AGING_DROP flush
modes, the root WBC EX locks will be dropped level by level during
flushing, to full support fsync(2) for these two flush mode, it
needs to first reopen the file from MDT when WBC EX lock revoking.

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I5c086e5351c9e5dc28b33428d31f0841abb9e194