LU-13008 wbc: inode limits for WBC caching
On a client, WBC borrows the design and implementation from
Linux/tmpfs a lot. It uses a virtual in-memory subtree to
represent a directory entirely caching on a client.
All inodes and directory entries (dentry for short) are stored in
memory and managed by Linux VFS layer, which is a common VFS
layer data structure with its own private data.
When create a file under a directory protected by the EX WBC
lock, the client only creates the corresponding in-memory inode
and dentry, then pins the dentry in the dentry cache by adding
the reference to the dentry object.
When unlink the file, put the corresponding reference. Thus, the
in-memory dentry and inode will be released when the last
reference to the dentry object is deleted.
Also similar to Linux/tmpfs, file data is directly written into
and pinned the page caches

To prevent from exhausting all virtual memory on a client, MemFS
used by WBC on the client side should allow an administrator to
specify a maximum upper bound for the caching size in two aspects:
- The maximum number of inodes;
- Page cache size for caching file data;

This patch implements the inode limits for WBC caching.
At the beginning, the root WBC dirtory is marked with Complete(C)
state.
All newly created files under the protection of the corresponding
root WBC EX lock is also marked with Complete(C) state together
with the state Reserved(E) state if consumed and reserved the
inode in MemFS successfully;
Otherwise, it must decomplete the parent directory. All IOs under
the parent directory must be blocked until decompleting is
finished. During decompleting, it must flush all its children to
MDT and then unmask the Complete(C) from the parent inode.
At this time, it needs to reclaim the reserved inodes: unmasked
the Reserved(E) state, dput() the corresponding dentry and
reclaim the inode count in MemFS.
For the directory not in Complete(C) state, all IOs under it must
issue and execute on MDT.
The file removal in the Complete(C) state is also need to
unreserve and reclaim the used inode count.

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I451f2bd901a1a2d4e88f82fc98af972fe7e644c0