LU-13521 wbc: readdir() handling for a directory under WBC
Under the protection of the root EX WBC lock, the open()/close()
system call does not need to communicate with MDS, can be executed
locally in MemFS on the client.

However, Lustre is a stateful filesystem. Each open keeps a certain
state on the MDS. WBC feature should keep transparency for
applications. To achieve this goal, it must reopen the non-closed
files from MDS when the root EX WBC lock is revoking.

It needs to handle a readdir() call carefully once the directory
has been opened. The reasons is as follows:
- Currently the mechanism adopted by MemFS (tmpfs/ramfs) is to
  simply scan the in-memory children dentries of the directory in
  dcache linearly to fill the content returned to readdir call:
  ->dcache_readdir().
- While Lustre new readdir implementation is much complex. It does
  readdir in hash order and uses hash of a file name as a
  telldir/seekdir cookie and stored as the position of the file
  handle.

For a regular file, the file position is logical offset in the
file, will keep same after reopen the file from MDT.
However, for a directory, it needs to bridge two implementation of
readdir() call due to the different traversal addressing
mechanisms for dirents under a directory.

In this patch, it implements two strategies for readdir() call:
- Directly use kernel dcache_readdir() to read dentries from
  dcache linearly. It is only used for benchmark. It may rarely
  read repeated and inconsistent dirents in case of reopening the
  file due to the revocation of the root WBC EX lock on the root
  WBC directory.
- Try readdir from dcache firstly if the directory is small enough
  to read all entries in one blow; Otherwise, it will decomplete
  the directory first, which means that flush all children
  dentries to MDT and unmask Complete(C) flag from the directory.
  And then it will read the children dentries from MDT in hash
  order.

Fortunately, the default buffer allocated for getdents system call
in libc is 32768 (32K) bytes. For a file name entry with the
length of 14 bytes, it will occupy 24 bytes to fill the
linux_dirent64 data structure. The allocated buffer is large enough
to hold nearly 1024 entries.

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I47d992ca155427a21712077cd4a4a685431b9778