LU-13011 wbc: data on PCC (DOP) for lock drop flush mode
The memory size on a client is limited compared with the local
persistent storage. For a WBC subtree, the metadata can be
reasonably whole cached in MemFS. But the data for regular files
maybe grows up too large to cache on client-side MemFS.

PCC can be used as Client Persistent Caching for the data of
regular files under the protection of EX WBC locks.

This patch integrates Persistent Client Cache (PCC) with WBC and
implements Data on PCC for the lock drop flush mode.
When flush the dirty inode of a regular file, the client will
select one aviable PCC backend, and MDT just creates the metadata
object with HSM archived, exists, released state and selected HSM
archive ID on MDT.
The client delays to instantiate the PCC copy until commit the
dirty pages. At this time, it will create the PCC copy stub
according to the FID and then write each uncommitted dirty page to
the PCC copy. All these operations do not require interaction with
the servers.

The file data cached on PCC can defer resync to Lustre OSTs or
evict from PCC when it is nearly full.

Next step, we will instantiate PCC copy when the file is growing
too large.

Test-Parameters: clientdistro=el8.3 testlist=sanity-wbc,sanity-wbc
Test-Parameters: testlist=sanity-wbc,sanity-wbc
Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: Ic10ca3e1455ce3f1e5d0c1c27f7d92a2b3200734