LU-13044 wbc: subtree removal for keep flush mode
This patch implements subtree removal strategy for (lazy or aging)
lock keep flush mode.

This file removal policy is similar to the delay removal policy
by FID(s).
The difference is that: when the parent directory is in C state
and a directory under this parent directory is to be deleted,
there is no need to splice two lists, just add the directroy entry
into the removed file list of the parent directory.
Since the client has exclusive access to the direcoty, it
indicates that the whole subtree has been removed from the file
system.
When the file is being deleted in the background flusher thread,
only the FID of the root removed node of the subtree is passed
to the server.
After received the removal request, the server is charged for the
subtree removal.
The server can also move the directory into PENDING directory just
like what we do for orphan object. The server can replay to the
client immediately, and then remove the subtree asynchronous on
background.

Signed-off-by: Qian Yingjin <qian@ddn.com>
Change-Id: I507f26cf96495e7c3c5db2daf144f4d7c8760f9d